graph: Add `api_version` module
